fastapi/requirements.txt
cryptography==38.0.1
fastapi==0.87.0
pydantic==1.8.2
python-dotenv==0.21.0
SQLAlchemy==1.4.31
uvicorn==0.19.0
psycopg2-binary==2.9.5